[TWINFACES-944] refactor: removed history to separate entitie
