+++ src/entities/history/api/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-fetch-by-id";
