+++ src/entities/history/api/index.ts
@@ -1,0 +1,3 @@
+export * from "./api-service";
+export * from "./hooks";
+export * from "./types";
