+++ src/entities/history/index.ts
@@ -1,0 +1,2 @@
+export * from "./api";
+export * from "./libs";
