+++ src/entities/twin/libs/helpers.ts
@@ -1,14 +1,5 @@
 import { DataListOptionV1 } from "@/entities/datalist-option";
-import { TwinClassField_DETAILED } from "@/entities/twin-class-field";
-import {
-  HistoryV1,
-  History_DETAILED,
-  Twin,
-  TwinTagManageV1,
-  Twin_DETAILED,
-} from "@/entities/twin/server";
-import { User } from "@/entities/user";
-import { RelatedObjects } from "@/shared/api";
+import { Twin, TwinTagManageV1 } from "@/entities/twin/server";
 import {
   isFalsy,
   isObject,
@@ -76,30 +67,3 @@ export function transformToTwinTags(
     { existingTags: [], newTags: [] }
   );
 }
-
-export const hydrateTwinHistoryFromMap = (
-  dto: HistoryV1,
-  relatedObjects?: RelatedObjects
-): History_DETAILED => {
-  const hydrated: History_DETAILED = Object.assign({}, dto) as History_DETAILED;
-
-  if (dto.machineUserId && relatedObjects?.userMap) {
-    hydrated.machineUser = relatedObjects.userMap[dto.machineUserId] as User;
-  }
-
-  if (dto.actorUserId && relatedObjects?.userMap) {
-    hydrated.actorUser = relatedObjects.userMap[dto.actorUserId] as User;
-  }
-
-  if (dto.twinClassFieldId && relatedObjects?.twinClassFieldMap) {
-    hydrated.twinClassField = relatedObjects.twinClassFieldMap[
-      dto.twinClassFieldId
-    ] as TwinClassField_DETAILED;
-  }
-
-  if (dto.twinId && relatedObjects?.twinMap) {
-    hydrated.twin = relatedObjects.twinMap[dto.twinId] as Twin_DETAILED;
-  }
-
-  return hydrated;
-};
