+++ src/entities/twin/libs/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./use-filters";
-export * from "./use-filters-history";
 export * from "./use-select-adapter";
 export * from "./use-twin-form-fields";
